refactor(snackbar): consolidate utility functions and enhance logging with package name prefixes
